+++ src/treg/api.py
@@ -146,6 +146,7 @@ from .routers.auth import (
     _cli_states,
     _client_ip,
     _find_or_create_user,
+    _live_invite_by_email_token,
     _login_page_html,
     _norm_pair_code,
     _orgs_brief,
@@ -160,6 +161,8 @@ from .routers.auth import (
     auth_github_callback,
     auth_google,
     auth_google_callback,
+    auth_invite_signin,
+    auth_invite_signin_confirm,
     auth_logout,
     auth_me,
     login_page,
@@ -661,113 +664,7 @@ router.routes.extend(auth_routes.session_router.routes)
 router.routes.extend(auth_routes.email_router.routes)
 
 
-async def _live_invite_by_email_token(db: AsyncSession, t: str) -> Invite | None:
-    """Resolve an emailed invite-link token to a live invite: pending, unexpired, unconsumed
-    (email_token_hash is nulled on first use), and not pointing at a platform-locked org."""
-    t = (t or "").strip()
-    if not t:
-        return None
-    invite = (await db.execute(select(Invite).where(Invite.email_token_hash == crypto.hash_token(t)))
-              ).scalar_one_or_none()
-    if (invite is None or invite.status != "pending"
-            or (invite.expires_at is not None and _as_naive(invite.expires_at) < _utcnow_naive())):
-        return None
-    org = await db.get(Org, invite.org_id)
-    if org is None or org.suspended:
-        return None
-    return invite
-
-
-@app.get("/auth/invite-signin")
-async def auth_invite_signin(
-    request: Request, code: str = "", t: str = "",
-    treg_session: str = Cookie(default=""), db: AsyncSession = Depends(get_session),
-):
-    """Landing for an invite email link. Two secrets, two very different trust levels:
-
-    `t` (email_token) exists ONLY in the emailed link — possession proves inbox access, the same bar
-    as the emailed OTP — so it may sign the invitee in. But not on this GET: corporate mail scanners
-    (Outlook SafeLinks etc.) prefetch GET links and would consume a one-time credential before the
-    human ever clicks. So the GET only renders a confirm page whose button POSTs the token back;
-    the POST below mints the session.
-
-    `code` (legacy + out-of-band) is also returned to the admin who created the invite, so it can
-    NEVER be an authentication factor — holding it lets you JOIN (POST /invites/accept), not log in.
-    Links carrying ?code= (emails sent before the split, or relayed by an admin) keep their old
-    behavior: validate and bounce to the SPA login with the email prefilled; the invitee proves the
-    email through a real door (OTP / GitHub / Google) and the invite auto-appears via /invites/mine.
-    An invalid/expired secret of either kind just lands on the site."""
-    from urllib.parse import quote
-    base = get_settings().public_url.rstrip("/")
-    if t:
-        invite = await _live_invite_by_email_token(db, t)
-        if invite is None:
-            return RedirectResponse("/?invite_expired=1", status_code=303)
-        org = await db.get(Org, invite.org_id)
-        # Already signed in as someone ELSE? Warn — continuing replaces that browser session.
-        switch_note = ""
-        uid = sess.read(treg_session)
-        if uid is not None:
-            current = await db.get(User, uid)
-            if current is not None and current.email != invite.email:
-                switch_note = (f"<p>You're currently signed in as <b>{_esc_html(current.email)}</b> — "
-                               f"continuing switches this browser to <b>{_esc_html(invite.email)}</b>.</p>")
-        return HTMLResponse(
-            f'{_AUTH_HEAD}<body><div class="wrap"><div class="card">'
-            f'<div class="logo">▚ tools-registry</div><div class="mark">👋</div>'
-            f'<h1>Join {_esc_html(org.name if org else "the team")}</h1>'
-            f'<p><b>{_esc_html(invite.invited_by or "A teammate")}</b> invited '
-            f'<b>{_esc_html(invite.email)}</b> as {_esc_html(invite.role)}.</p>{switch_note}'
-            f'<form method="post" action="/auth/invite-signin" style="margin-top:18px">'
-            f'<input type="hidden" name="t" value="{_esc_html(t.strip())}">'
-            f'<button type="submit" class="pbtn">'
-            f'Continue as {_esc_html(invite.email)} →</button></form>'
-            f"</div></div></body></html>"
-        )
-    c = (code or "").strip()
-    invite = (await db.execute(select(Invite).where(Invite.code_hash == crypto.hash_token(c)))
-              ).scalar_one_or_none() if c else None
-    if (invite is None or invite.status != "pending"
-            or (invite.expires_at is not None and _as_naive(invite.expires_at) < _utcnow_naive())):
-        return RedirectResponse("/?invite_expired=1", status_code=303)
-    # Code path: same redirect whether or not the email already has an account — the code is a
-    # convenience that prefills the sign-in email, never an authentication factor. A suspended
-    # account is caught at the real login door, the only place the code path can mint a session.
-    return RedirectResponse(f"/?invite={quote(invite.email)}", status_code=303)
-
-
-@app.post("/auth/invite-signin")
-async def auth_invite_signin_confirm(request: Request, db: AsyncSession = Depends(get_session)):
-    """The confirm page's POST: the emailed one-time token signs the invitee in. Mirrors the OTP
-    door (auth_email_verify) — find-or-create the user, refuse the suspended, set the session
-    cookie — because the trust source is identical: only the inbox saw this secret. The token is
-    consumed here (one-time) so a link floating in a forwarded thread can't be replayed; the invite
-    itself stays PENDING — acceptance happens in the dashboard, where a multi-team invitee can
-    accept several at once. Body is parsed by hand (urlencoded form) to avoid the python-multipart
-    dependency FastAPI's Form() would pull in."""
-    from urllib.parse import parse_qs
-    try:
-        form = parse_qs((await request.body()).decode("utf-8", "replace"))
-    except Exception:  # noqa: BLE001 — any junk body = no token
-        form = {}
-    t = (form.get("t", [""])[0] or "").strip()
-    invite = await _live_invite_by_email_token(db, t)
-    if invite is None:  # consumed / expired / revoked / suspended org → the SPA's expired banner
-        return RedirectResponse("/?invite_expired=1", status_code=303)
-    user = await _find_or_create_user(db, invite.email)  # first click = registration (user only, no auto org)
-    if user is None or user.suspended:  # a banned account may hold the link but must not get a session
-        return _auth_page("Account suspended", "This account has been suspended.", ok=False, status=403)
-    invite.email_token_hash = None  # consume: one sign-in per emailed link
-    db.add(invite)
-    await db.commit()
-    # A share-born invite lands on the shared page itself (the SPA auto-accepts + switches org);
-    # a plain invite lands on the dashboard with the accept banner, as before. `landing` was
-    # allowlist-validated at create time, so this can never redirect off-app.
-    dest = f"{invite.landing}?invite_org={invite.org_id}" if invite.landing else f"/?invite_org={invite.org_id}"
-    resp = RedirectResponse(dest, status_code=303)
-    resp.set_cookie(sess.COOKIE, sess.make(user.id, token_version=user.token_version), httponly=True,
-                    samesite="lax", secure=_is_https(request), max_age=sess.TTL_SECONDS)
-    return resp
+router.routes.extend(auth_routes.invite_router.routes)
 
 
 # Register the moved site routes at their original position.
